Travis: use stormpy docker containers

+++ travis/build-helper.sh
@@ -11,45 +11,17 @@ travis_fold() {
 
 # Helper for building and testing
 run() {
-  # Create virtual environment
-  virtualenv --python=$PYTHON prophesy-env
-  source prophesy-env/bin/activate
+  travis_fold start virtualenv
+  if [[ "$CONFIG" != *Stormpy* ]]
+  then
+    # Create virtual environment
+    virtualenv --python=$PYTHON venv
+  fi
+  # Activate virtual environment
+  source venv/bin/activate
   # Print version
   python --version
-
-  # Build pycarl
-  if [[ $CONFIG != *"Pip"* ]]; then
-    travis_fold start build_pycarl
-    git clone https://github.com/moves-rwth/pycarl.git
-    cd pycarl
-    case "$CONFIG" in
-    Debug*)
-      python setup.py build_ext --debug -j 1 develop
-      ;;
-    *)
-      python setup.py build_ext -j 1 develop
-      ;;
-    esac
-    travis_fold end build_pycarl
-    cd ..
-  fi
-
-  # Build stormpy
-  if [[ $CONFIG != *"Pip"* ]]; then
-    travis_fold start build_stormpy
-    git clone https://github.com/moves-rwth/stormpy.git
-    cd stormpy
-    case "$CONFIG" in
-    Debug*)
-      python setup.py build_ext --storm-dir /opt/storm/build/ --debug -j 1 develop
-      ;;
-    *)
-      python setup.py build_ext --storm-dir /opt/storm/build/ -j 1 develop
-      ;;
-    esac
-    travis_fold end build_stormpy
-    cd ..
-  fi
+  travis_fold end virtualenv
 
   # Build prophesy
   travis_fold start build_prophesy
@@ -57,29 +29,25 @@ run() {
   travis_fold end build_prophesy
 
   # Perform task
-  case $TASK in
-  Test)
+  if [[ "$TASK" == *Test* ]]
+  then
     # Install pytest
     pip install pytest
     # Run tests
     set +e
     python -m pytest tests
     python -m pytest scripts/tests
-    ;;
+  fi
 
-  Documentation)
+  if [[ "$TASK" == *Documentation* ]]
+  then
     # Generate documentation
     pip install sphinx sphinx_bootstrap_theme
     cd doc
     make html
     touch build/html/.nojekyll
     rm -r build/html/_sources
-    ;;
-
-  *)
-    echo "Unrecognized value of TASK: $TASK"
-    exit 1
-  esac
+  fi
 }
 
 run
